simplify queue
