--- packages/@eventual/core/src/queue/index.ts
@@ -1,2 +1,0 @@
-export * from "./queue.js";
-export * from "./fifo.js";
